refactor api routes, get auth working with mci-auth
