+++ .env.sample
@@ -1,4 +1,13 @@
+NODE_ENV=development
+
+AWS_ACCESS_KEY_ID=XXXXXXXXXX
+AWS_SECRET_ACCESS_KEY=xxxxxxxxxx
+AWS_REGION=us-east-1
+AWS_BUCKET=bucket-name
+
 EMAIL_DB=emails
 IMAGE_DB=images
 USER_DB=users
 COUCHDB_URL=http://localhost:5984/
+
+AUTH_URL=https://auth.example.com/authenticate/
