+++ package.json
@@ -33,10 +33,13 @@
     "dotenv": "^2.0.0",
     "express": "^4.13.4",
     "express-handlebars": "^3.0.0",
+    "express-session": "^1.15.1",
     "handlebars": "^4.0.5",
     "juice": "^4.0.2",
     "moment": "^2.17.1",
     "multer": "^1.2.1",
+    "passport": "^0.3.2",
+    "passport-local": "^1.0.0",
     "pg": "^6.1.0",
     "pouchdb-browser": "^6.1.1",
     "react": "^15.4.1",
